requirements.txt
fastapi==0.100.1
pydantic~=1.10.4
starlette~=0.27.0
passlib~=1.7.4
sqlalchemy~=2.0.0
alembic~=1.11.2
python-jose
